Add function to do math calculations
Created functions to take the user input for math operations and
properly use it to calculate two numbers. Allowed the user to chain
multiple operations for multiple numbers. Also added functionality to
the decimal button and clear button. These changes are all needed to
make the calculator app perform like an actual calculator.